requirements.txt
PySide6
requests
python-dotenv
pandas
openpyxl
